- code compile checks - storage unit tests logic updated

+++ include/countly/storage_module_base.hpp
@@ -55,7 +55,7 @@ public:
    * Returns the count of stored requests.
    * @return count of stored requests
    */
-  virtual int RQCount() = 0;
+  virtual long long RQCount() = 0;
 
   /**
    * Delete all stored requests.
